Added getData function to populate data from server and changed data from int based to double based

+++ Assessments/CourseData.h
@@ -5,20 +5,20 @@
 // Holds all the data for a course
 class CourseData {
 public:
-    int courseNum;                      // Course number
-    int year;                           // Year
-    std::string semester;               // Semester
-    std::string code;                   // Code ex. CSCI
-    std::string name;                   // Name of course
-    std::string section;                // Section
-    std::string prof;                   // Professor's name
-    std::vector<std::string> comps;     // Courses competencies
-    std::vector<int> mastery;           // Points to get mastery in competency
-    std::vector<std::vector<int>> data; // 2D vector to store student data
-    std::vector<double> average;        // Averages in each competency
-    std::vector<double> median;         // Medians in each competency
-    std::vector<double> percent;        // Percent passed in each competency
-    std::vector<double> deviation;      // Deviation in each competency
+    int courseNum;                          // Course number
+    int year;                               // Year
+    std::string semester;                   // Semester
+    std::string code;                       // Code ex. CSCI
+    std::string name;                       // Name of course
+    std::string section;                    // Section
+    std::string prof;                       // Professor's name
+    std::vector<std::string> comps;         // Courses competencies
+    std::vector<double> mastery;            // Points to get mastery in competency
+    std::vector<std::vector<double>> data;  // 2D vector to store student data
+    std::vector<double> average;            // Averages in each competency
+    std::vector<double> median;             // Medians in each competency
+    std::vector<double> percent;            // Percent passed in each competency
+    std::vector<double> deviation;          // Deviation in each competency
 
     void calculate();   // Calculate average, median, median, and deviation for course
 
@@ -30,8 +30,8 @@ public:
     void setSection(std::string sectionIn) { section = sectionIn; }             // Sets section
     void setProf(std::string profIn) { prof = profIn; }                         // Sets professors name
     void addComps(std::string compsIn) { comps.push_back(compsIn); }            // Adds a competency
-    void addMastery(int masteryIn) { mastery.push_back(masteryIn); }            // Adds a mastery
-    void addData(std::vector<int> dataIn) { data.push_back(dataIn); }           // Adds new student data
+    void addMastery(double masteryIn) { mastery.push_back(masteryIn); }         // Adds a mastery
+    void addData(std::vector<double> dataIn) { data.push_back(dataIn); }        // Adds new student data
     void addAverage(double averageIn) { average.push_back(averageIn); }         // Adds a average
     void addMedian(double medianIn) { median.push_back(medianIn); }             // Adds a median
     void addPercent(double percentIn) { percent.push_back(percentIn); }         // Adds a percent
@@ -45,8 +45,8 @@ public:
     const std::string* getSection() { return &section; }                // Returns section by reference
     const std::string* getProf() { return &prof; }                      // Returns professors name by reference
     const std::vector<std::string>* getComps() { return &comps; }       // Returns competencies by reference
-    const std::vector<int>* getMastery() { return &mastery; }           // Returns masteries by reference
-    const std::vector<std::vector<int>>* getData() { return &data; }    // Returns data by reference
+    const std::vector<double>* getMastery() { return &mastery; }        // Returns masteries by reference
+    const std::vector<std::vector<double>>* getData() { return &data; } // Returns data by reference
     const std::vector<double>* getAverage() { return &average; }        // Returns averages by reference
     const std::vector<double>* getMedian() { return &median; }          // Returns medians by reference
     const std::vector<double>* getPercent() { return &percent; }        // Returns percents by reference
