+++ Assessments/Util.h
@@ -4,5 +4,5 @@
 
 namespace util {
     std::string curlRequest(std::string url);   // Get data from online servers from given URL
-    std::vector<std::vector<double>> calcData(std::vector<std::vector<int>>);   // Calculates the course data
+    std::vector<std::vector<double>> calcData(std::vector<std::vector<double>>);    // Calculates the course data
 }
